refactor: rationalizes settings screen
