+++ core/l10n/src/androidMain/res/values/strings.xml
@@ -205,9 +205,9 @@
     <string name="settings_post_layout_compact">Compact</string>
     <string name="settings_post_layout_full">Full</string>
     <string name="settings_section_appearance">Look and feel</string>
-    <string name="settings_section_behaviour">Behaviour</string>
+    <string name="settings_advanced">Advanced settings</string>
     <string name="settings_section_debug">Debug</string>
-    <string name="settings_section_feed">Posts and comments</string>
+    <string name="settings_section_general">General</string>
     <string name="settings_section_nsfw">NSFW</string>
     <string name="settings_theme_black">Pure black</string>
     <string name="settings_theme_dark">Dark</string>
