+++ feature/settings/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/feature/settings/main/SettingsScreen.kt
@@ -13,7 +13,6 @@ import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.AdminPanelSettings
 import androidx.compose.material.icons.filled.BugReport
-import androidx.compose.material.icons.filled.Dashboard
 import androidx.compose.material.icons.filled.Explicit
 import androidx.compose.material.icons.filled.Menu
 import androidx.compose.material.icons.filled.SettingsApplications
@@ -36,13 +35,10 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.ColorFilter
 import androidx.compose.ui.input.nestedscroll.nestedScroll
-import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.platform.LocalUriHandler
 import androidx.compose.ui.text.buildAnnotatedString
-import androidx.compose.ui.unit.toSize
 import cafe.adriel.voyager.core.screen.Screen
 import cafe.adriel.voyager.koin.getScreenModel
-import com.github.diegoberaldin.raccoonforlemmy.core.appearance.data.UiBarTheme
 import com.github.diegoberaldin.raccoonforlemmy.core.appearance.data.toReadableName
 import com.github.diegoberaldin.raccoonforlemmy.core.appearance.theme.Spacing
 import com.github.diegoberaldin.raccoonforlemmy.core.architecture.bindToLifecycle
@@ -50,12 +46,8 @@ import com.github.diegoberaldin.raccoonforlemmy.core.commonui.lemmyui.SettingsHe
 import com.github.diegoberaldin.raccoonforlemmy.core.commonui.lemmyui.SettingsRow
 import com.github.diegoberaldin.raccoonforlemmy.core.commonui.lemmyui.SettingsSwitchRow
 import com.github.diegoberaldin.raccoonforlemmy.core.commonui.lemmyui.handleUrl
-import com.github.diegoberaldin.raccoonforlemmy.core.commonui.modals.BarThemeBottomSheet
-import com.github.diegoberaldin.raccoonforlemmy.core.commonui.modals.DurationBottomSheet
-import com.github.diegoberaldin.raccoonforlemmy.core.commonui.modals.InboxTypeSheet
 import com.github.diegoberaldin.raccoonforlemmy.core.commonui.modals.LanguageBottomSheet
 import com.github.diegoberaldin.raccoonforlemmy.core.commonui.modals.ListingTypeBottomSheet
-import com.github.diegoberaldin.raccoonforlemmy.core.commonui.modals.SliderBottomSheet
 import com.github.diegoberaldin.raccoonforlemmy.core.commonui.modals.SortBottomSheet
 import com.github.diegoberaldin.raccoonforlemmy.core.commonui.modals.ThemeBottomSheet
 import com.github.diegoberaldin.raccoonforlemmy.core.l10n.LocalXmlStrings
@@ -67,12 +59,11 @@ import com.github.diegoberaldin.raccoonforlemmy.core.notifications.di.getNotific
 import com.github.diegoberaldin.raccoonforlemmy.core.utils.compose.onClick
 import com.github.diegoberaldin.raccoonforlemmy.core.utils.compose.rememberCallback
 import com.github.diegoberaldin.raccoonforlemmy.core.utils.compose.rememberCallbackArgs
-import com.github.diegoberaldin.raccoonforlemmy.core.utils.datetime.getPrettyDuration
 import com.github.diegoberaldin.raccoonforlemmy.core.utils.toLanguageFlag
 import com.github.diegoberaldin.raccoonforlemmy.core.utils.toLanguageName
-import com.github.diegoberaldin.raccoonforlemmy.core.utils.toLocalDp
 import com.github.diegoberaldin.raccoonforlemmy.domain.lemmy.data.toInt
 import com.github.diegoberaldin.raccoonforlemmy.domain.lemmy.data.toReadableName
+import com.github.diegoberaldin.raccoonforlemmy.feature.settings.advanced.AdvancedSettingsScreen
 import com.github.diegoberaldin.raccoonforlemmy.feature.settings.colors.SettingsColorAndFontScreen
 import com.github.diegoberaldin.raccoonforlemmy.unit.about.AboutDialog
 import com.github.diegoberaldin.raccoonforlemmy.unit.accountsettings.AccountSettingsScreen
@@ -83,7 +74,6 @@ import com.github.diegoberaldin.raccoonforlemmy.unit.web.WebViewScreen
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.launch
-import kotlin.math.roundToInt
 
 class SettingsScreen : Screen {
 
@@ -118,11 +108,8 @@ class SettingsScreen : Screen {
             }.launchIn(this)
         }
 
-        var screenWidth by remember { mutableStateOf(0f) }
         Scaffold(
-            modifier = Modifier.onGloballyPositioned {
-                screenWidth = it.size.toSize().width
-            }.padding(Spacing.xxs),
+            modifier = Modifier.padding(Spacing.xs),
             topBar = {
                 TopAppBar(
                     scrollBehavior = scrollBehavior,
@@ -190,56 +177,6 @@ class SettingsScreen : Screen {
                         },
                     )
 
-                    // navigation bar titles
-                    SettingsSwitchRow(
-                        title = LocalXmlStrings.current.settingsNavigationBarTitlesVisible,
-                        value = uiState.navBarTitlesVisible,
-                        onValueChanged = rememberCallbackArgs(model) { value ->
-                            model.reduce(
-                                SettingsMviModel.Intent.ChangeNavBarTitlesVisible(value)
-                            )
-                        },
-                    )
-
-                    // edge to edge
-                    SettingsSwitchRow(
-                        title = LocalXmlStrings.current.settingsEdgeToEdge,
-                        value = uiState.edgeToEdge,
-                        onValueChanged = rememberCallbackArgs(model) { value ->
-                            model.reduce(
-                                SettingsMviModel.Intent.ChangeEdgeToEdge(value)
-                            )
-                        },
-                    )
-
-                    // system bar theme
-                    if (uiState.edgeToEdge) {
-                        val barThemeName = if (uiState.opaqueSystemBars) {
-                            UiBarTheme.Opaque.toReadableName()
-                        } else {
-                            UiBarTheme.Transparent.toReadableName()
-                        }
-                        SettingsRow(
-                            title = LocalXmlStrings.current.settingsBarTheme,
-                            value = barThemeName,
-                            onTap = rememberCallback {
-                                val sheet = BarThemeBottomSheet()
-                                navigationCoordinator.showBottomSheet(sheet)
-                            },
-                        )
-                    }
-
-                    // bottom navigation hiding
-                    SettingsSwitchRow(
-                        title = LocalXmlStrings.current.settingsHideNavigationBar,
-                        value = uiState.hideNavigationBarWhileScrolling,
-                        onValueChanged = rememberCallbackArgs(model) { value ->
-                            model.reduce(
-                                SettingsMviModel.Intent.ChangeHideNavigationBarWhileScrolling(value)
-                            )
-                        },
-                    )
-
                     // colors and fonts
                     SettingsRow(
                         title = LocalXmlStrings.current.settingsColorsAndFonts,
@@ -249,19 +186,18 @@ class SettingsScreen : Screen {
                         }
                     )
 
-                    SettingsHeader(
-                        icon = Icons.Default.Dashboard,
-                        title = LocalXmlStrings.current.settingsSectionFeed,
-                    )
-
-
-                    // colors and fonts
+                    // content view configuration
                     SettingsRow(
                         title = LocalXmlStrings.current.settingsConfigureContent,
                         disclosureIndicator = true,
                         onTap = rememberCallback {
                             navigationCoordinator.pushScreen(ConfigureContentViewScreen())
                         }
+                    )
+
+                    SettingsHeader(
+                        icon = Icons.Default.SettingsApplications,
+                        title = LocalXmlStrings.current.settingsSectionGeneral,
                     )
 
                     // default listing type
@@ -307,94 +243,6 @@ class SettingsScreen : Screen {
                     )
 
                     if (uiState.isLogged) {
-                        // default inbox type
-                        SettingsRow(
-                            title = LocalXmlStrings.current.settingsDefaultInboxType,
-                            value = if (uiState.defaultInboxUnreadOnly) {
-                                LocalXmlStrings.current.inboxListingTypeUnread
-                            } else {
-                                LocalXmlStrings.current.inboxListingTypeAll
-                            },
-                            onTap = rememberCallback {
-                                val sheet = InboxTypeSheet()
-                                navigationCoordinator.showBottomSheet(sheet)
-                            },
-                        )
-                    }
-
-                    SettingsHeader(
-                        icon = Icons.Default.SettingsApplications,
-                        title = LocalXmlStrings.current.settingsSectionBehaviour,
-                    )
-
-                    // auto-expand comments
-                    SettingsSwitchRow(
-                        title = LocalXmlStrings.current.settingsAutoExpandComments,
-                        value = uiState.autoExpandComments,
-                        onValueChanged = rememberCallbackArgs(model) { value ->
-                            model.reduce(
-                                SettingsMviModel.Intent.ChangeAutoExpandComments(value)
-                            )
-                        },
-                    )
-
-                    // infinite scrolling
-                    SettingsSwitchRow(
-                        title = LocalXmlStrings.current.settingsInfiniteScrollDisabled,
-                        value = uiState.infiniteScrollDisabled,
-                        onValueChanged = rememberCallbackArgs(model) { value ->
-                            model.reduce(
-                                SettingsMviModel.Intent.ChangeInfiniteScrollDisabled(value)
-                            )
-                        },
-                    )
-
-                    if (uiState.isLogged) {
-                        // mark as read while scrolling
-                        SettingsSwitchRow(
-                            title = LocalXmlStrings.current.settingsMarkAsReadWhileScrolling,
-                            value = uiState.markAsReadWhileScrolling,
-                            onValueChanged = rememberCallbackArgs(model) { value ->
-                                model.reduce(
-                                    SettingsMviModel.Intent.ChangeMarkAsReadWhileScrolling(value)
-                                )
-                            },
-                        )
-                    }
-
-                    // zombie mode interval
-                    SettingsRow(
-                        title = LocalXmlStrings.current.settingsZombieModeInterval,
-                        value = uiState.zombieModeInterval.getPrettyDuration(
-                            secondsLabel = LocalXmlStrings.current.postSecondShort,
-                            minutesLabel = LocalXmlStrings.current.postMinuteShort,
-                            hoursLabel = LocalXmlStrings.current.homeSortTypeTop6Hours,
-                        ),
-                        onTap = rememberCallback {
-                            val sheet = DurationBottomSheet()
-                            navigationCoordinator.showBottomSheet(sheet)
-                        },
-                    )
-
-                    // zombie scroll amount
-                    SettingsRow(
-                        title = LocalXmlStrings.current.settingsZombieModeScrollAmount,
-                        value = buildString {
-                            val pt = uiState.zombieModeScrollAmount.toLocalDp().value.roundToInt()
-                            append(pt)
-                            append(LocalXmlStrings.current.settingsPointsShort)
-                        },
-                        onTap = rememberCallback {
-                            val sheet = SliderBottomSheet(
-                                min = 0f,
-                                max = screenWidth,
-                                initial = uiState.zombieModeScrollAmount,
-                            )
-                            navigationCoordinator.showBottomSheet(sheet)
-                        },
-                    )
-
-                    if (uiState.isLogged) {
                         // swipe actions
                         SettingsSwitchRow(
                             title = LocalXmlStrings.current.settingsEnableSwipeActions,
@@ -413,17 +261,6 @@ class SettingsScreen : Screen {
                                 navigationCoordinator.pushScreen(screen)
                             },
                         )
-
-                        // double tap
-                        SettingsSwitchRow(
-                            title = LocalXmlStrings.current.settingsEnableDoubleTap,
-                            value = uiState.enableDoubleTapAction,
-                            onValueChanged = rememberCallbackArgs(model) { value ->
-                                model.reduce(
-                                    SettingsMviModel.Intent.ChangeEnableDoubleTapAction(value)
-                                )
-                            },
-                        )
                     }
 
                     // URL open
@@ -437,25 +274,13 @@ class SettingsScreen : Screen {
                         },
                     )
 
-                    // image loading
-                    SettingsSwitchRow(
-                        title = LocalXmlStrings.current.settingsAutoLoadImages,
-                        value = uiState.autoLoadImages,
-                        onValueChanged = rememberCallbackArgs(model) { value ->
-                            model.reduce(
-                                SettingsMviModel.Intent.ChangeAutoLoadImages(value)
-                            )
-                        },
-                    )
-
-                    // search posts only in title
-                    SettingsSwitchRow(
-                        title = LocalXmlStrings.current.settingsSearchPostsTitleOnly,
-                        value = uiState.searchPostTitleOnly,
-                        onValueChanged = rememberCallbackArgs(model) { value ->
-                            model.reduce(
-                                SettingsMviModel.Intent.ChangeSearchPostTitleOnly(value)
-                            )
+                    // advanced settings
+                    SettingsRow(
+                        title = LocalXmlStrings.current.settingsAdvanced,
+                        disclosureIndicator = true,
+                        onTap = rememberCallback {
+                            val screen = AdvancedSettingsScreen()
+                            navigationCoordinator.pushScreen(screen)
                         },
                     )
